requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
beautifulsoup4==4.8.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
Django==3.1.3
django-crispy-forms==1.10.0
django-queryset-csv==1.1.0
django-telegrambot==1.0.1
gunicorn==20.0.4
idna==2.10
lxml==4.6.1
parse-hh-data==0.1.14
psycopg2==2.8.6
pycparser==2.20
pydantic==1.7.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.0
pytz==2020.4
random-user-agent==1.0.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tornado==6.1
tqdm==4.44.1
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.25.8
